refactor, move flow dialog components
