+++ frontend/src/components/Workspace/FlowChart/Dialog/AlgorithmOutputDialog.tsx
